+++ src/cancelable/cancelify.js
@@ -1,8 +1,9 @@
 import { curryN } from 'ramda';
+import addFantasyLandInterface from "./addFantasyLandInterface.js";
 
 const
-	cancelify = fn => curryN(fn.length, (...args) =>
-		(res, rej) => {
+	cancelify = fn => curryN(fn.length, (...args) => {
+		const cancelable = (res, rej) => {
 			let
 				resolveInner = res,
 				rejectInner = rej;
@@ -13,7 +14,11 @@ const
 				resolveInner = () => undefined;
 				rejectInner = () => undefined;
 			};
-		}
-	);
+		};
+		
+		addFantasyLandInterface(cancelable);
+		
+		return cancelable;
+	});
 
 export default cancelify;