+++ src/transformations/baconObservableToCancelable.js
@@ -1,16 +1,23 @@
 import * as Bacon from 'baconjs';
+import addFantasyLandInterface from "../cancelable/addFantasyLandInterface.js";
 
 const
-	baconObservableToCancelable = observable => (res, rej) =>
-		observable.last().subscribe(evt => {
-			if (evt.hasValue) {
-				res(evt.value);
-			}
-			else if (evt.isError) {
-				rej(evt.error);
-			}
-			return Bacon.noMore;
-		});
+	baconObservableToCancelable = observable => {
+		const cancelable = (res, rej) =>
+			observable.last().subscribe(evt => {
+				if (evt.hasValue) {
+					res(evt.value);
+				}
+				else if (evt.isError) {
+					rej(evt.error);
+				}
+				return Bacon.noMore;
+			});
+		
+		addFantasyLandInterface(cancelable);
+		
+		return cancelable;
+	}
 
 export default baconObservableToCancelable;